Add JPA

+++ driver-service/src/main/java/com/driver/service/DriverService.java
@@ -4,5 +4,5 @@ import com.driver.web.model.LocationRequest;
 import com.driver.web.model.LocationResponse;
 
 public interface DriverService {
-    public LocationResponse updateLocation(Long id, LocationRequest locationUpdateRequest);
+    public LocationResponse updateLocation(int id, LocationRequest locationUpdateRequest);
 }
